Commit message: Add user-drag and text-size-adjust properties for better cross-browser support Fixes issues with text selection and dragging on mobile devices, improving overall user experience. updated main.scss and _variables.scss to include new spacing and typography variables for enhanced design consistency across the application. These changes will allow for more flexible and responsive layouts, as well as improved readability on various screen sizes. Additionally, updated index.html to ensure the new styles are properly applied. Loading Code...

+++ market_service.go
@@ -37,7 +37,7 @@ func (l *Lobby) handleTradeShares(env *Envelope) {
 	targetWallet = strings.ToLower(targetWallet)
 	targetStats := l.leaderboard[targetWallet]
 
-	basePrice := float64((targetStats.Wins * 10) + (float64(targetStats.Reputation) / 2.0) + 100.0)
+	basePrice := float64((targetStats.Wins * 10) + int(float64(targetStats.Reputation)/2.0) + 100.0)
 	for _, rumor := range l.rumors {
 		if strings.EqualFold(rumor.TargetWallet, targetWallet) && time.Now().Before(rumor.ExpiresAt) {
 			basePrice *= rumor.Strength
